user management

+++ MedRecordManager/Controllers/UserAdminController.cs
@@ -1,16 +1,19 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using MedRecordManager.Areas.Identity.Pages.Account;
 using MedRecordManager.Data;
 using MedRecordManager.Extension;
 using MedRecordManager.Models;
 using MedRecordManager.Models.UserRecord;
+using MedRecordManager.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using UrgentCareData;
 using UrgentCareData.Models;
@@ -24,12 +27,14 @@ namespace MedRecordManager.Controllers
         private readonly ApplicationDbContext _userContext;
         private readonly UrgentCareContext _urgentCareContext;
         private readonly UserManager<ApplicationUser> _userManager;
-        public UserAdminController(RoleManager<IdentityRole> roleManager, UserManager<ApplicationUser> userManager, UrgentCareContext _urgentCareContext, ApplicationDbContext _userContext)
+        private readonly IViewRenderService _viewRenderService;
+        public UserAdminController(RoleManager<IdentityRole> roleManager, UserManager<ApplicationUser> userManager, UrgentCareContext _urgentCareContext, ApplicationDbContext _userContext, IViewRenderService viewRenderService)
         {
             _roleManager = roleManager;
             _userManager = userManager;
             this._urgentCareContext = _urgentCareContext;
             this._userContext = _userContext;
+            _viewRenderService = viewRenderService;
         }
 
 
@@ -132,11 +137,23 @@ namespace MedRecordManager.Controllers
                 {
                     UserId = user.Id,
                     FirstName = user.FirstName,
-                    Email = user.Email
+                    Email = user.Email,
+                    LastName = user.LastName
                 };
 
                 var roles = await _userManager.GetRolesAsync(user);
                 thisUser.Roles = string.Join("</br>", roles);
+
+                var companies = await _urgentCareContext.UserCompany.Where(x => x.UserId == user.Id).ToListAsync();
+                thisUser.Companies = string.Join("</br>", companies);
+
+                var offices = await _urgentCareContext.UserOfficeKey.Where(x => x.UserId == user.Id).ToListAsync();
+                thisUser.OfficeKeys = string.Join("</br>", offices);
+
+                var clinics = await _urgentCareContext.UserClinic.Where(x => x.UserId == user.Id).ToListAsync();
+                thisUser.Clinics = string.Join("</br>", clinics);
+
+
                 records.Add(thisUser);
             }
 
@@ -155,34 +172,48 @@ namespace MedRecordManager.Controllers
         [HttpDelete]
         public async Task<IActionResult> DeleteUser(string userName)
         {
+            var currentuser = this.User; 
+            
             var user = await _userManager.FindByIdAsync(userName);
-            var result = await _userManager.DeleteAsync(user);
-            if (result.Succeeded)
+            if(user != null)
             {
-                return Json(new { success = true});
+                if (currentuser.Identity.Name != user.Email)
+                {
+                    var result = await _userManager.DeleteAsync(user);
+                    if (result.Succeeded)
+                    {
+                        return Json(new { success = true });
+                    }
+                }
+                else
+                {
+                    return Json(new { success = false, message="You can not delete yourself while you are logged in. "});
+                }
             }
-            return Json(new { success = false });
+            
+            return Json(new { success = false , message="Either this user does not exisits or there is some thing wrong when tried to delete."});
         }
-
         public async Task<IActionResult> ManageRole()
         {
             var roles = await _roleManager.Roles.ToListAsync();
             return View("ManageRole",roles);
         }
 
-
         public IActionResult ManageCompany()
-        {
-           
+        {         
             return View("ManageCompany");
         }
-
         public IActionResult ManageClinic()
         {
 
             return View("ManageClinic");
         }
 
+        public IActionResult ManageOffice()
+        {
+
+            return View("ManageOffice");
+        }
 
         public async Task<IActionResult> GetUserCompanies(int? page, int? limit)
         {
@@ -202,14 +233,18 @@ namespace MedRecordManager.Controllers
             return Json(new { records, total });
         }
 
-        
-
         public async Task<IActionResult> getUserClinics(int? page, int? limit)
         {
 
             var records = new List<ClinicProfile>();
-
-            records = await _urgentCareContext.ClinicProfile.Where(x=>x.OfficeKey != null).ToListAsync();
+            if(User.IsInRole("SuperAdmin"))
+            {
+                records = await _urgentCareContext.ClinicProfile.Where(x => x.OfficeKey != null).ToListAsync();
+            }
+           else if(User.IsInRole("CompanyAdmin"))
+            {
+                
+            }
             records.ForEach(x => x.ClinicName = x.ClinicId);
             var total = records.Count();
 
@@ -223,5 +258,171 @@ namespace MedRecordManager.Controllers
             return Json(new { records, total });
         }
 
+        public async Task<IActionResult> getUserOffices(int? page, int? limit)
+        {
+
+            var records = new List<ClinicProfile>();
+
+            records = _urgentCareContext.ClinicProfile.Where(x => x.OfficeKey != null).DistinctBy(x=>x.OfficeKey).ToList();
+           
+            var total = records.Count();
+
+            if (page.HasValue && limit.HasValue)
+            {
+                var start = (page.Value - 1) * limit.Value;
+                records = records.Skip(start).Take(limit.Value).ToList();
+
+
+            }
+            return Json(new { records, total });
+        }
+
+        [HttpGet]
+        public async Task<IActionResult> AssignPermission(string userId)
+        {
+            var currentUser = new UserVm();
+            var user = await _userManager.FindByIdAsync(userId);
+            var aRoles = await _userManager.GetRolesAsync(user);
+            var sRoles = await _roleManager.Roles.ToListAsync();
+           
+            if (user != null)
+            {
+
+                currentUser.Email = user.Email;
+                currentUser.FirstName = user.FirstName;
+                currentUser.LastName = user.LastName;
+                currentUser.UserId = user.Id;
+               
+            }
+
+            var isSuperAdmin = await _userManager.IsInRoleAsync(user, "SuperAdmin");
+            if (!isSuperAdmin)
+            {
+                currentUser.AvaliableRoles = await _roleManager.Roles.Where(x=> !aRoles.Contains(x.Name)).Select(x => new SelectListItem
+                {
+                    Text = x.Name,
+                    Value = x.Id
+
+                }).ToListAsync();
+            }
+
+            var roleNames =await  _userManager.GetRolesAsync(user);
+            currentUser.AssignedRoles = await _roleManager.Roles.Where(x => roleNames.Contains(x.Name)).Select(x => new SelectListItem
+            {
+                Text = x.Name,
+                Value = x.Id
+
+            }).ToListAsync();
+
+            currentUser.AvaliableComps = await _urgentCareContext.CompanyProfile.Select(x=> new SelectListItem
+            {
+                Text= x.CompanyName,
+                Value = x.Id.ToString()
+            }).ToListAsync();
+
+            currentUser.AssignedComps = await _urgentCareContext.UserCompany.Include(x=>x.Company).Where(x=>x.UserId == currentUser.UserId).Select(x => new SelectListItem
+            {
+                Text = x.Company.CompanyName,
+                Value = x.CompanyId.ToString()
+
+            }).ToListAsync();
+
+            currentUser.AvaliableOffices =  _urgentCareContext.ClinicProfile.Where(x=>x.OfficeKey !=null).DistinctBy(x=>x.OfficeKey).Select(x => new SelectListItem
+            {
+                Text = x.OfficeKey.ToString(),
+                Value = x.OfficeKey.ToString()
+            }).ToList();
+
+            currentUser.AssignedOffices = await _urgentCareContext.UserOfficeKey.Where(x => x.UserId == currentUser.UserId).Select(x => new SelectListItem
+            {
+                Text = x.OfficeKey.ToString(),
+                Value = x.OfficeKey.ToString()
+
+            }).ToListAsync();
+
+            currentUser.AvailableClinics = _urgentCareContext.ClinicProfile.DistinctBy(x => x.ClinicId).Select(x => new SelectListItem
+            {
+                Text = x.ClinicId,
+                Value = x.ClinicId
+            }).ToList();
+
+            currentUser.AssignedClinics = await _urgentCareContext.UserClinic.Where(x => x.UserId == currentUser.UserId).Select(x => new SelectListItem
+            {
+                Text = x.ClinicId,
+                Value = x.ClinicId
+
+            }).ToListAsync();
+
+            return PartialView("~/Views/UserAdmin/UserPermission.cshtml", currentUser);
+        }
+
+        [HttpPost]
+        public async Task<IActionResult> UpdateUserPerm(string userId, string firstName, string lastName, List<string> assignedRoles, List<string> assignedComps, List<string> assignedOffices, List<string> assignedClinics)
+        {
+            try
+            {
+                var appUser = await _userManager.FindByIdAsync(userId);
+                var currentRoles = await _userManager.GetRolesAsync(appUser);
+
+                await _userManager.RemoveFromRolesAsync(appUser, currentRoles);
+                await _userManager.AddToRolesAsync(appUser, assignedRoles);
+
+                var oldComps= await _urgentCareContext.UserCompany.Where(x => x.UserId == userId).ToListAsync();
+                _urgentCareContext.UserCompany.RemoveRange(oldComps);
+                foreach (var comp in assignedComps)
+                {
+                    _urgentCareContext.UserCompany.Add(new UserCompany { 
+                        UserId = userId,
+                        CompanyId = int.Parse(comp)
+                    });
+                }
+
+                var oldOffices = await  _urgentCareContext.UserOfficeKey.Where(x => x.UserId == userId).ToListAsync();
+                _urgentCareContext.UserOfficeKey.RemoveRange(oldOffices);
+                foreach (var office in assignedOffices)
+                {
+                    _urgentCareContext.UserOfficeKey.Add(new UserOfficeKey
+                    {
+                        UserId = userId,
+                        OfficeKey = int.Parse(office)
+                    });
+                }
+
+                var oldClinics = await _urgentCareContext.UserClinic.Where(x => x.UserId == userId).ToListAsync();
+                _urgentCareContext.UserClinic.RemoveRange(oldClinics);
+                foreach (var clinic in assignedClinics)
+                {
+                    _urgentCareContext.UserClinic.Add(new UserClinic
+                    {
+                        UserId = userId,
+                        ClinicId = clinic
+                    });
+                }
+
+                _urgentCareContext.SaveChanges();
+            }
+            catch
+            {
+                return Json(new { success = false });
+            }
+
+            return RedirectToAction("AssignPermission", new { userId });
+        }
+        //public async Task<IActionResult> getAvaliableRoles()
+        //{
+        //    var roles = new List<SelectListItem>();
+        //    var user = await _userManager.GetUserAsync(User);
+        //    var assinedroles =await _userManager.GetRolesAsync(user);
+        //    if(User.IsInRole("SuperAdmin"))
+        //    {
+        //        roles=  await _roleManager.Roles.Select(x=> new SelectListItem { 
+        //            Text = x.Name,
+        //            Value = x.Id
+
+        //        }).ToListAsync();
+        //    }
+
+        //return Json(roles);
+        //}
     }
 }
